# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.BankAccount:
    properties:
      _id:
        type: string
      account_holder:
        type: string
      currency:
        type: string
      iban:
        type: string
      name:
        type: string
      user_id:
        type: string
    type: object
...
  requests.BankAccountCreate:
    properties:
      account_holder:
        type: string
      currency:
        type: string
      iban:
        type: string
      name:
        type: string
    required:
    - account_holder
    - currency
    - iban
    - name
    type: object
  requests.BankAccountUpdate:
    properties:
      account_holder:
        type: string
      currency:
        type: string
      iban:
        type: string
      id:
        type: string
      name:
        type: string
    required:
    - id
    type: object
...
  requests.ID:
    properties:
      id:
        type: string
    required:
    - id
    type: object
...
  responses.TransactionTotal:
    properties:
      currency:
        type: string
      total_transaction:
        type: number
    type: object
...
host: rocky-reaches-65250.herokuapp.com
info:
  contact:
    email: mrntlu@gmail.com
    name: Burak Fidan
  description: REST Api of Kantan.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://rocky-reaches-65250.herokuapp.com/terms/
  title: Kantan Investment Manager API
  version: "1.0"
paths:
...
  /ba:
    delete:
      consumes:
      - application/json
      description: Deletes bank account by id
      parameters:
      - description: ID
        in: body
        name: ID
        required: true
        schema:
          $ref: '#/definitions/requests.ID'
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete bank account by bank account id
      tags:
      - bankaccount
    get:
      consumes:
      - application/json
      description: Returns bank accounts by user id
      parameters:
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.BankAccount'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get Bank Accounts by User ID
      tags:
      - bankaccount
    post:
      consumes:
      - application/json
      description: Creates bank account
      parameters:
      - description: Bank Account Create
        in: body
        name: bankaccount
        required: true
        schema:
          $ref: '#/definitions/requests.BankAccountCreate'
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.BankAccount'
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create BankAccount
      tags:
      - bankaccount
    put:
      consumes:
      - application/json
      description: Updates bank account
      parameters:
      - description: BankAccount Update
        in: body
        name: bankaccountupdate
        required: true
        schema:
          $ref: '#/definitions/requests.BankAccountUpdate'
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BankAccount'
        "400":
          description: Bad Request
          schema:
            type: string
        "403":
          description: Unauthorized update
          schema:
            type: string
        "404":
          description: Couldn't find user
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update Bank Account
      tags:
      - bankaccount
  /ba/all:
    delete:
      consumes:
      - application/json
      description: Deletes all bank accounts by user id
      parameters:
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete all bank accounts by user id
      tags:
      - bankaccount
  /ba/stats:
    get:
      consumes:
      - application/json
      description: Returns bank accounts stats
      parameters:
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.TransactionTotal'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get Bank Accounts Stats
      tags:
      - bankaccount
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"